Rust is the bees knees. Got people. Got some fake auth. Good times.

+++ settlement-management-sheet/client/src/app/thunks/glossary/subtypes/groups/deleteSubTypeGroupThunk.ts
@@ -3,19 +3,25 @@ import { AppThunk } from '@/app/thunks/glossaryThunks.js';
 import { RootState } from '@/app/store.js';
 import { showSnackbar } from '@/app/slice/snackbarSlice.js';
 import { addDirtyKeypath } from '@/app/slice/dirtySlice.js';
-import { addSubTypeGroup, deleteGroup } from '@/app/slice/subTypeSlice.js';
-import deleteSubTypeGroupService from '@/services/glossary/subTypes/deleteSubTypeGroupService.js';
+import {
+  addSubTypeGroup,
+  deleteGroup,
+  SubTypeGroup,
+} from '@/app/slice/subTypeSlice.js';
 import { cloneDeep } from 'lodash';
 import subTypeCommands from '@/app/commands/subtype.ts';
+import getSubTypeStateAtKey from '@/utility/dataTransformation/getSubTypeStateAtKey.ts';
+import checkAdmin from '@/utility/security/checkAdmin.ts';
 
 export function deleteSubTypeGroupThunk({
   groupId,
 }: {
   groupId: string;
 }): AppThunk {
-  return async (dispatch: ThunkDispatch<RootState, unknown, any>, getState) => {
-    const state = getState();
-    const group = cloneDeep(state.subType.groups.edit[groupId]);
+  return async (dispatch: ThunkDispatch<RootState, unknown, any>) => {
+    const group = cloneDeep(
+      getSubTypeStateAtKey<SubTypeGroup>('groups', groupId)
+    );
     if (!group) {
       dispatch(
         showSnackbar({
@@ -26,25 +32,20 @@ export function deleteSubTypeGroupThunk({
       );
       return;
     }
-    try {
-      // await deleteSubTypeGroupService({ groupId });
-      await subTypeCommands.deleteSubTypeGroup({ id: groupId });
 
+    if (!checkAdmin(group.system)) return;
+
+    try {
       dispatch(
         deleteGroup({
           groupId,
+          system: group.system,
         })
       );
 
-      dispatch(
-        addDirtyKeypath({
-          scope: 'subType',
-          id: groupId,
-          keypath: `${groupId}`,
-        })
-      );
+      await subTypeCommands.deleteSubTypeGroup({ id: groupId });
     } catch (error) {
-      dispatch(addSubTypeGroup({ groups: [group] }));
+      dispatch(addSubTypeGroup({ groups: [group], system: group.system }));
       dispatch(
         showSnackbar({
           message: 'Error removing group. Try again later.',
